Address PR #29 review: split boards, move codegen to library, PCI child model
Restore qemu-aarch64 to its original LinuxBoot+ATF config and create
a dedicated qemu-aarch64-uefi board for the CrabEFI UEFI payload.
This keeps the general-purpose board unchanged while the UEFI config
gets its own RON file with PCI, GIC, and Bochs display.

Move the ~120-line EFI memory map construction from codegen-generated
inline code to fstart_crabefi::build_efi_memory_map(). The codegen
now emits a single function call instead of building the RAM split
logic inline. This makes the memory map logic testable, readable,
and reusable across boards.

Fix FDT sourcing to be generic: use src_dtb_addr from payload config
or fstart_platform::boot_dtb_addr() instead of hardcoding RAM base.
This removes the QEMU-virt-specific assumption and works across
platforms.

Convert BochsDisplay from Device to BusDevice with PCI bus parent.
The driver now receives its BARs from the parent PciRootBus's config
space accessors via new_on_bus() instead of duplicating the ECAM base
and scanning PCI config space independently. The board RON nests
bochs0 under pci0's children with just device:function addressing.

Fix clippy warnings in fstart-driver-pci-ecam (unnecessary casts)
and fstart-codegen (unused variable).

+++ crates/fstart-driver-pci-ecam/src/lib.rs
@@ -236,7 +236,7 @@ impl PciEcam {
         if original & 1 == 1 {
             // I/O BAR
             let mask = sized | 0x3;
-            let size = ((!mask) as u32).wrapping_add(1) as u64;
+            let size = (!mask).wrapping_add(1) as u64;
             return (
                 BarInfo {
                     bar_type: BarType::Io,
@@ -256,7 +256,7 @@ impl PciEcam {
             0 => {
                 // 32-bit
                 let mask = sized | 0xF;
-                let size = ((!mask) as u32).wrapping_add(1) as u64;
+                let size = (!mask).wrapping_add(1) as u64;
                 (
                     BarInfo {
                         bar_type: BarType::Memory32,
